+++ Launcher/GameLauncher.cs
@@ -40,6 +40,8 @@ namespace F8Framework.Launcher
             FF8.Download = ModuleCenter.CreateModule<DownloadManager>();
             FF8.LogWriter = ModuleCenter.CreateModule<F8LogWriter>();
 
+            yield return new WaitForEndOfFrame();
+            
             StartGame();
             yield break;
         }