+++ Tests/Localization/Example/Scripts/LanguageCycler.cs
@@ -43,6 +43,8 @@ namespace F8Framework.Tests
 			FF8.Download = ModuleCenter.CreateModule<DownloadManager>();
 			FF8.LogWriter = ModuleCenter.CreateModule<F8LogWriter>();
 			
+			yield return new WaitForEndOfFrame();
+			
 			StoryCanvas.SetActive(true);
 			UICanvas.SetActive(true);
 			yield break;